Sudoku Game
This is another sudoku game . Where user can solve the sudoku game and can also sudoku game as well as level wise . There are 3 levels : Easy , Medium and Hard. And when any user want to play sudoku he/she have to insert his/her name ,gender . The data of the user will be stored in exell formate in local computer.